Rename *.gdi/*.cdi file only when necessary
Closes #2

+++ dc-card-maker.sh
@@ -208,9 +208,11 @@ while read GAME; do
         # to target directory, add the game to the game list
         echo "Generating $ARCHIVE_FILE"
         echo "$GAME" > "$DIR_NAME/$ARCHIVE_FILE"
-        echo "Renaming disc file \"$DISC_FILE\" to \"disc.$TYPE\""
         # If moving files goes wrong abort immediately
-        mv "$DIR_NAME/$DISC_FILE" "$DIR_NAME/disc.$TYPE" || exit
+        if [[ ! -e "$DIR_NAME/disc.$TYPE" ]]; then
+            echo "Renaming disc file \"$DISC_FILE\" to \"disc.$TYPE\""
+            mv "$DIR_NAME/$DISC_FILE" "$DIR_NAME/disc.$TYPE" || exit
+        fi
         echo "Moving game to target directory"
         mv "$DIR_NAME" "$TARGET_DIR" || exit
         (( INDEX++ ))
